activities.jsx refactor
