+++ pages/activities/index.js
@@ -1,0 +1,2 @@
+// TODO: cleanup indirection
+module.exports = require('./ActivitiesPage.jsx');
